Added Viewport
Added code for a viewport, which allows for the screen to keep the same perspective with a centered view that changes when the window is resized.

Also fixed the player movement to include the delta.

+++ LeviathanPC/stdafx.h
@@ -12,8 +12,8 @@
 #include "Vector.h"
 
 //Reciporicals of \sqrt{2}
-#define SQRT_2 (float) 0.70710678118
-#define _SQRT_2 (float) -0.70710678118
+#define SQRT_2 0.70710678118f
+#define _SQRT_2 -0.70710678118f
 
 //Eight directon enum
 enum EightDirection {
